requirements.txt
djangorestframework
pyyaml
django-cors-headers
django-phonenumber-field[phonenumbers]
Django==4.1.4
markdown 
django-filter
psycopg2-binary
djangorestframework-simplejwt
PyJWT==2.1.0
django-dotenv
drf-access-policy